making things pretty
