--- Server_Rocket/lib/MySD/MySD.h
@@ -1,8 +1,0 @@
-#include <Arduino.h>
-#include <MyCommon.h>
-#include <SD.h>
-
-void setup_sd();
-
-void log_data(struct dataPoint *telemetry);
-
